Refactor: ConfirmProvider로 모달 관련 코드 통합 및 경로 수정
